[handlers] extract common, profile, dose, and reporting modules
